CNAME
These two entries tell GitHub to redirect any requests to antohneo.github.io to the domain specified in the CNAME file. There are two entries to ensure that antonioa.com will also redirect to www.antonioa.com. If you prefer to have a site without the www prefix, you can switch the order of the domains.